Task 8.4: Эмодзи (клиентская часть)

+++ packages/server/app/services/reaction.service.ts
@@ -7,25 +7,14 @@ export class ReactionService {
   }
 
   async setReaction(commentId: number, reactionId: number, author: string) {
-    await CommentReaction.create({
-      commentId: commentId,
-      author: author,
-      reactionId: reactionId
-    });
+    await CommentReaction.create({ commentId, author, reactionId });
   }
 
-  async deleteReaction(commentId: number, reactionId: number) {
-    await CommentReaction.destroy({
-      where: {
-        commentId: commentId,
-        reactionId: reactionId
-      }
-    });
+  async deleteReaction(commentId: number, reactionId: number, author: string) {
+    await CommentReaction.destroy({ where: { commentId, author, reactionId } });
   }
 
-  async addReaction(emoji: string) {
-    await Reaction.create({
-      emoji: emoji
-    });
+  async addManyReactions(emojis: string[]) {
+    await Reaction.bulkCreate(emojis.map((emoji) => ({ emoji })));
   }
 }
